Modify Course to use published boolean instead of publish event
